build baseAssert class + fix basicSteps

+++ src/Betsson.OnlineWallets.APITests/Models/BalanceResponse.cs
@@ -2,5 +2,5 @@
 
 public class BalanceResponse
 {
-    public decimal Amount { get; set; }
+    public double Amount { get; set; }
 }
